requirements.txt
torch
nltk
scikit-learn
matplotlib
seaborn
numpy
speechrecognition
google-generativeai
flask
flask-cors
python-dotenv
